Refactor wireless views into separate views for each hardware component
