+++ autosportlabs/racecapture/views/configuration/rcp/configview.py
@@ -136,25 +136,18 @@ class ConfigView(Screen):
             return tree.add_node(LinkedTreeViewLabel(text=text, is_open=True, no_selection=True))
 
         def show_node(node):
-            try:
-                view = node.view
-                if not view:
-                    view = node.view_builder()
-                    self.configViews.append(view)
-                    view.bind(on_config_modified=self.on_config_modified)
-                    node.view = view
-                    if self.loaded:
-                        if self.rc_config:
-                            view.dispatch('on_config_updated', self.rc_config)
-                        if self.track_manager:
-                            view.dispatch('on_tracks_updated', self.track_manager)
-                Clock.schedule_once(lambda dt: self.ids.content.add_widget(view))
-
-
-            except Exception as detail:
-                alertPopup('Error', 'Error loading screen ' + str(node.text) + ':\n\n' + str(detail))
-                Logger.error("ConfigView: Error selecting configuration view " + str(node.text))
-                Logger.error("ConfigView: {}".format(detail))
+            view = node.view
+            if not view:
+                view = node.view_builder()
+                self.configViews.append(view)
+                view.bind(on_config_modified=self.on_config_modified)
+                node.view = view
+                if self.loaded:
+                    if self.rc_config:
+                        view.dispatch('on_config_updated', self.rc_config)
+                    if self.track_manager:
+                        view.dispatch('on_tracks_updated', self.track_manager)
+            Clock.schedule_once(lambda dt: self.ids.content.add_widget(view))
 
         def on_select_node(instance, value):
             # ensure that any keyboard is released
@@ -198,7 +191,7 @@ class ConfigView(Screen):
 
         attach_node('CAN Bus', None, lambda: CANConfigView())
         attach_node('OBDII', None, lambda: OBD2ChannelsView(channels=runtime_channels, base_dir=self.base_dir))
-        attach_node('Wireless', None, lambda: WirelessConfigView(base_dir=self.base_dir))
+        attach_node('Wireless', None, lambda: WirelessConfigView(self.base_dir))
 
         if self.rc_config.capabilities.has_cellular:
             attach_node('Telemetry', None, lambda: TelemetryConfigView())
