+++ autosportlabs/racecapture/views/configuration/rcp/wireless/bluetoothconfigview.kv
@@ -1,0 +1,17 @@
+#:kivy 1.9.1
+
+<BluetoothConfigView>
+    id: bluetooth
+    cols: 1
+    spacing: [0, dp(20)]
+    row_default_height: dp(40)
+    size_hint: [1, None]
+    height: self.minimum_height
+    HSeparator:
+        size_hint_y: 0.5
+        text: 'Bluetooth'
+    SettingsView:
+        id: bt_enable
+        label_text: 'Bluetooth'
+        help_text: 'If the Bluetooth module is connected, enable it here'
+        size_hint_y: 1
